% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angularmomentum.R
\name{angularmomentum}
\alias{angularmomentum}
\title{Compute the total angular momentum of a particle distribution}
\usage{
angularmomentum(m, x, v, x0 = NULL, v0 = NULL)
}
\arguments{
\item{m}{n-vector or scalar of particle masses. For specific angular momentum computation, set \code{m=1}.}

\item{x}{\eqn{n\times3} matrix of particle positions.}

\item{v}{\eqn{n\times3} matrix of particle velocities.}

\item{x0}{Optional 3-vector specifying the origin about which the
angular momentum is computed. If \code{NULL} (default), the centre of
mass is used.}

\item{v0}{Optional 3-vector specifying the reference velocity. If
\code{NULL} (default), the centre-of-mass velocity is used.}
}
\value{
A numeric vector of length three containing the Cartesian components of the
total angular momentum,
\deqn{\mathbf{J}=\sum_i m_i\,(\mathbf{x}_i-\mathbf{x}_0)\times
(\mathbf{v}_i-\mathbf{v}_0).}

If \code{m} is \code{NULL}, a vector of zeros is returned.
}
\description{
Computes the total angular momentum of a set of particles about their
centre of mass (or about a user-specified origin).
}
\examples{
m = c(1, 2)
x = rbind(c(1, 0, 0), c(0, 1, 0))
v = rbind(c(0, 1, 0), c(-1, 0, 0))
angularmomentum(m, x, v)

}
