% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtree.R
\name{buildtree}
\alias{buildtree}
\title{Build self-similar merger tree}
\usage{
buildtree(m = c(1, 1), nlevels = 2)
}
\arguments{
\item{m}{vector of relative masses (normalization is irrelevant)}

\item{nlevels}{number of progenitor levels}
}
\description{
Builds a self-similar rooted directional tree (e.g. a halo merger tree) to be displayed with \code{\link{plottree}}.
}
\examples{

## build and show a self-similar binary tree with mass ratio 2:1
tree = buildtree(m=c(2,1), nlevels=4)
plottree(tree)

}
\seealso{
\code{\link{plottree}}
}
\author{
Danail Obreschkow
}
