% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car2pol.R
\name{car2pol}
\alias{car2pol}
\title{Cartesian to polar/cylindrical coordinate conversion}
\usage{
car2pol(x)
}
\arguments{
\item{x}{2/3-element or n-by-2/3 matrix representing the Cartesian components (x,y)/(x,y,z)  of n two/three-dimensional vectors}
}
\value{
Returns a 2/3-element vector or a n-by-2/3 element matrix representing the polar/cylindrical coordinates (r,phi)/(r,phi,z), where phi=0...2*pi is the azimuth measured positively from the x-axis.
}
\description{
Convert 2D/3D Cartesian to polar/cylindrical coordinates
}
\seealso{
\code{\link{pol2car}}
}
\author{
Danail Obreschkow
}
