% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear memory}
\usage{
clear(environment = TRUE, plots = TRUE, console = TRUE)
}
\arguments{
\item{environment}{logical flag indicating whether to clear the environment}

\item{plots}{logical flag indicating whether to clear the RStudio plots}

\item{console}{logical flag indicating whether to clear the RStudio console}
}
\value{
None
}
\description{
Clears the shared environment, plots and console
}
\author{
Danail Obreschkow
}
