% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooltools-package.R
\docType{package}
\name{cooltools}
\alias{cooltools}
\alias{cooltools-package}
\title{Routines for scientific research}
\description{
This package contains a useful collection of routines for scientific research, grown organically through practical experience. These routines span across numerical analysis, graphics, statistics and system functions.
}
\author{
Danail Obreschkow
}
