% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshift.R
\name{cshift}
\alias{cshift}
\title{Circularly shift each dimension of an array}
\usage{
cshift(x, s)
}
\arguments{
\item{x}{vector or array (up to rank 5)}

\item{s}{scalar, if x is a vector, or a vector of length matching the rank of x, if x is an array}
}
\value{
Returns a vector or array of the same shape as x.
}
\description{
Circulates each dimension of an array. This routine is identical to \code{\link[pracma]{circshift}}, but works with arrays up to 5 dimensions.
}
\author{
Danail Obreschkow
}
