% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst.R
\name{cst}
\alias{cst}
\title{Scientific constants}
\usage{
cst
}
\value{
None
}
\description{
The list \code{cst} contains useful scientific constants in SI units, mainly for astrophysics.
}
\examples{
# print all the constants to console
for (i in seq(length(cst))) cat(sprintf('\%6s = \%.12e\n',names(cst)[i],cst[i]))

}
\author{
Danail Obreschkow
}
