% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft.R
\name{dft}
\alias{dft}
\title{Discrete Fourier Transform}
\usage{
dft(f, inverse = F)
}
\arguments{
\item{f}{real or complex array containing the values to be transformed.}

\item{inverse}{logical flag; if TRUE the inverse Fourier Transfrom is performed.}
}
\value{
Returns an array of the same shape as \code{f}, containing the (inverse) Fourier Transform.
}
\description{
Discrete Fourier Transform (DFT) with largest modes at the centre in k-space and normalized such that dft(dft(f),inverse)=f. This is the discretization scheme described in  Appendix D of Obreschkow et al. 2013, ApJ 762. Relies on \code{\link[stats]{fft}}.
}
\examples{

## DFT of a 2D normal Gaussian function
n = 30
f = array(0,c(n,n))
for (i in seq(n)) {
  for (j in seq(n)) f[i,j] = exp(-(i-6)^2/4-(j-8)^2/2-(i-6)*(j-8)/2)
}
qplot(xlim=c(0,2.1),ylim=c(0,1.1),asp=1)
rasterImage(rasterflip(f),0,0,1,1,interpolate=FALSE)
g = dft(f)
img = array(hsv((pracma::angle(g)/2/pi)\%\%1,1,abs(g)/max(abs(g))),c(n,n))
rasterImage(rasterflip(img),1.1,0,2.1,1,interpolate=FALSE)
text(0.5,1,'Input function f',pos=3)
text(1.6,1,'DFT(f)',pos=3)

}
\author{
Danail Obreschkow
}
