% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftgrid.R
\name{dftgrid}
\alias{dftgrid}
\title{Produce coordinates for Discrete Fourier Transform}
\usage{
dftgrid(N, L)
}
\arguments{
\item{N}{number of divisions along one dimension used in direct and Fourier space}

\item{L}{side-length (=period) of the data long one dimension}
}
\value{
Returns a list with the vectors x and k, the corresponding steps dx and dk, and the input values N and L.
}
\description{
Produces the coordinates in direct space (x) and Fourier space (k), corresponding to the Discrete Fourier Transform \code{dft} of this package.
}
\author{
Danail Obreschkow
}
