% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.env.R
\name{.cooltools.env}
\alias{.cooltools.env}
\title{Package environment}
\usage{
.cooltools.env
}
\value{
None
}
\description{
Environment used to store global variables in the package, e.g. used for subplot routine.
}
\author{
Danail Obreschkow
}
