% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{dpqr}
\alias{dpqr}
\title{Generate the d/p/q/r family for an arbitrary distribution}
\usage{
dpqr(fun, min, max)
}
\arguments{
\item{fun}{A non-negative, vectorised function proportional to a probability density. It does not need to be normalised.}

\item{min, max}{Lower and upper bounds of the support. The density is assumed to be zero outside this interval.}
}
\value{
Returns a list with the following functions:
\item{d(x)}{Probability density function (PDF), i.e. a normalised version of \code{fun} on the domain \code{[min,max]}.}
\item{p(x)}{Cumulative distribution function (CDF).}
\item{q(p)}{Quantile function.}
\item{r(n)}{Generates \code{n} random samples from the distribution.}
}
\description{
Generates the standard \code{d}, \code{p}, \code{q}, and \code{r}
functions associated with an arbitrary one-dimensional probability density
function, analogous to \code{dnorm}/\code{pnorm}/\code{qnorm}/\code{rnorm}
and related families.
}
\examples{
f = function(x) sin(x)
dist = dpqr(f, 0, pi)

x = dist$r(1000)
hist(x, probability = TRUE)
curve(dist$d(x), add = TRUE)

}
\seealso{
\code{\link{rng}}, \code{\link{contourlevel}}
}
\author{
Danail Obreschkow
}
