% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errlines.R
\name{errlines}
\alias{errlines}
\title{Draw a line with uncertainty regions}
\usage{
errlines(
  x,
  y,
  errp,
  errn = errp,
  col = "black",
  alpha = 0.5,
  smooth = FALSE,
  df = NULL,
  ...
)
}
\arguments{
\item{x}{vector of x-coordinates}

\item{y}{vector of y-coordinates}

\item{errp}{vector of y-errors in the positive direction (upwards)}

\item{errn}{vector of y-errors in the negative direction (downwards)}

\item{col}{color of the line}

\item{alpha}{transparency of the uncertainty region}

\item{smooth}{logical flag indicating whether the line should be smoothed}

\item{df}{df (=degrees of freedom) parameter of \code{\link[stats]{smooth.spline}} function}

\item{...}{additional parameters used by \code{\link[graphics]{lines}}}
}
\value{
None
}
\description{
Draw a line with uncertainty regions
}
\author{
Danail Obreschkow
}
