% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortune_teller.R
\name{fortune_teller}
\alias{fortune_teller}
\title{A function to tell your fortune}
\usage{
fortune_teller(number)
}
\arguments{
\item{number}{An integer between 1 and 4.}
}
\description{
A function to tell your fortune
}
\examples{
fortune_teller(1)
}
\keyword{luck}
