% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Compute gradient}
\usage{
gradient(f, circular = FALSE)
}
\arguments{
\item{f}{vector or array of rank 2 or 3}

\item{circular}{logical scalar or vector specifying whether periodic boundaries are used along each dimension}
}
\value{
returns a list with the different components of the gradient; each component has the same dimension as f.
}
\description{
Evaluates the gradient of a vector of array of rank 2 or 3
}
\author{
Danail Obreschkow
}
