% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Last element of a vector}
\usage{
last(x, n = 1)
}
\arguments{
\item{x}{vector}

\item{n}{optional integer specifying the n-th element from the end to be returned}
}
\value{
scalar of the same type as \code{x}
}
\description{
Returns the last element of a vector or the n-th element counting from the end of a vector.
}
\author{
Danail Obreschkow
}
