% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightness.R
\name{lightness}
\alias{lightness}
\title{Change lightness of a color}
\usage{
lightness(col, light = 0.5)
}
\arguments{
\item{col}{is a color or vector/array of colors, specified as text (e.g. 'purple') or 7/9-character (e.g. '#A020F0')}

\item{light}{lightness value, according to a HSL scheme, between 0 and 1 or a vector/array thereof}
}
\value{
Returns a 9-character color or vector/array of 9-character colors.
}
\description{
Change lightness of a color
}
\examples{

# Generate different lightnesses of the same color
plot(runif(50),runif(50),pch=20,cex=10,col=lightness('purple',runif(50)))

}
\seealso{
\code{\link{transparent}}
}
\author{
Danail Obreschkow
}
