% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midseq.R
\name{midseq}
\alias{midseq}
\title{Mid-points of regular grid}
\usage{
midseq(min, max, n = 1)
}
\arguments{
\item{min}{left boundary of first bin}

\item{max}{right boundary of last bin}

\item{n}{number of bins}
}
\value{
vector of mid points
}
\description{
compute the mid-point positions of a one-dimensional regular grid of n equal intervals (=bins)
}
\author{
Danail Obreschkow
}
