% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.R
\name{mutual}
\alias{mutual}
\title{Mutual information of two random variables}
\usage{
mutual(x, y = NULL, b = exp(1), ...)
}
\arguments{
\item{x}{either of the following: (1) an m-by-n matrix representing the 2D probability mass function of two random variables X and Y; all elements must be non-negative; the normalization is irrelevant. (2) an n-vector of sampled x-values; in this case y must be specified.}

\item{y}{optional vector of sampled y-values (only used if \code{x} is a vector of x-values).}

\item{b}{base of the logarithm in mutual information I(X,Y). Default is e.}

\item{...}{additional optional arguments (n, xlim, ylim) for the gridding routine \code{\link{griddata2}}, if x and y vectors are given.}
}
\value{
Returns a list of items:
\item{I}{standard mutual information I(X,Y).}
\item{N}{normalized mutual information I(X,Y)/sqrt(H(X)*H(Y)), where H is the Shannon information entropy.}
\item{pmf}{normalized, gridded probability mass function.}
}
\description{
Computes the mutual information of two random variables X and Y, given their 2D represented in a matrix.
}
\author{
Danail Obreschkow
}
