% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2jpg.R
\name{pdf2jpg}
\alias{pdf2jpg}
\title{Convert pdf to jpg}
\usage{
pdf2jpg(
  pdf.filename,
  jpg.filename,
  quality = 100,
  background = "white",
  dim = c(1600, 1200),
  remove.pdf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pdf.filename}{filename of pdf input file}

\item{jpg.filename}{filename of jpeg output file}

\item{quality}{quality of jpeg compression from 1 (worst) to 100 (best)}

\item{background}{color of background}

\item{dim}{size in pixels of the jpeg-image}

\item{remove.pdf}{logical flag. If TRUE, the pdf file is deleted after the conversion.}

\item{verbose}{logical flag to turn on/off console statements.}
}
\value{
None
}
\description{
Calls the console "convert" function to convert a pdf-file into a jpeg-image. Requires "convert" to be installed already.
}
\author{
Danail Obreschkow
}
