% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2car.R
\name{pol2car}
\alias{pol2car}
\title{Polar/cylindrical to Cartesian coordinate conversion}
\usage{
pol2car(x)
}
\arguments{
\item{x}{2/3-element or n-by-2/3 matrix representing the polar/cylindrical coordinates (r,phi)/(r,phi,z), where phi=0...2*pi is the azimuth measured positively from the x-axis.}
}
\value{
Returns a 2/3-element vector or a n-by-2/3 element matrix representing the Cartesian coordinates (x,y)/(x,y,z).
}
\description{
Convert polar/cylindrical coordinates to 2D/3D Cartesian coordinates
}
\seealso{
\code{\link{car2pol}}
}
\author{
Danail Obreschkow
}
