% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Show progress while timer in running}
\usage{
progress(txt = NULL)
}
\arguments{
\item{txt}{custom text to be displayed. If not provided, the time since calling tick() is displayed.}
}
\value{
None
}
\description{
Show progress while a timer is running following a call of tick().
}
\examples{

tick('Test')
Sys.sleep(.1)
for (i in seq(3)) {
  progress(i)
  Sys.sleep(.1)
}
tock()

}
\seealso{
\code{\link{tick}} \code{\link{tock}} \code{\link{stop}}
}
\author{
Danail Obreschkow
}
