% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3D.R
\name{r3D}
\alias{r3D}
\title{Generate randomly oriented vectors in 3D}
\usage{
r3D(n = 1, r = c(0, 1), azimuth = c(0, 2 * pi), polarangle = c(0,
  pi))
}
\arguments{
\item{n}{number of random vectors to be generated}

\item{r}{2-vector specifying the range of radii}

\item{azimuth}{2-vector specifying the range of azimuth angles}

\item{polarangle}{2-vector specifying the range of polar angles}
}
\value{
Returns an n-by-3 array of n vectors.
}
\description{
Generate randomly oriented vectors in 3D, following an isotropic distribution (optionally truncated to a region).
}
\examples{
## draw 20 unit vectors on a sphere
x = r3D(20,r=c(1,1))
print(rowSums(x^2))

}
\author{
Danail Obreschkow
}
