% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterflip.R
\name{rasterflip}
\alias{rasterflip}
\title{Flip array to be displayed with rasterImage()}
\usage{
rasterflip(A)
}
\arguments{
\item{A}{n-by-m or n-by-m-by-k array of an image}
}
\description{
Flips the raster image A to be displayed with rasterImage, such that the first index runs
}
\examples{
col = transparent('purple',0.5)

}
\author{
Danail Obreschkow
}
