% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation2.R
\name{rotation2}
\alias{rotation2}
\title{2D rotation matrix}
\usage{
rotation2(angle)
}
\arguments{
\item{angle}{rotation angle in radians (counter-clockwise)}
}
\value{
Returns a 2-by-2 anti-symmetric rotation matrix
}
\description{
Compute a 2D rotation matrix given a rotation angle
}
\seealso{
\code{\link{rotation3}}
}
\author{
Danail Obreschkow
}
