% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsphere.R
\name{rsphere}
\alias{rsphere}
\title{Generate random unit vectors in 3D}
\usage{
rsphere(n, azimuth = c(0, 2 * pi), polarangle = c(0, pi))
}
\arguments{
\item{n}{number of random vectors to be generated}

\item{azimuth}{2-vector specifying the range of azimuth angles}

\item{polarangle}{2-vector specifying the range of polar angles}
}
\value{
Returns an n-by-3 array of n vectors.
}
\description{
Generate random unit vectors in 3D, following an isotropic distribution (optionally truncated to a region).
}
\author{
Danail Obreschkow
}
