% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtree.R
\name{showtree}
\alias{showtree}
\title{Display merger tree}
\usage{
showtree(tree = NULL, normalize.mass = TRUE,
  monotonic.growth = FALSE, sort = "", simplify = FALSE,
  draw.edges = TRUE, draw.vertices = FALSE,
  draw.leaf.vertices = draw.vertices, draw.box = TRUE,
  root.length = 0.05, leaf.length = 0, style = 3, scale = 1,
  gamma = 1, min = 0.03, vertex.size = 1.5, margin = 0.05,
  add = FALSE, xlim = c(0, 1), ylim = c(0, 1), xrange = xlim,
  yrange = ylim, col = rainbow(100), pdf.filename = NULL,
  pdf.size = 5)
}
\arguments{
\item{tree}{list of three vectors of equal lengths, specifying the merger tree. Each element refers to one vertex (halo) and the root vertex must be listed first. \code{mass} contains the masses of the vertices (halos). \code{descendant} contains the descendant indices; the first element must be 0, indicating that the root has no decendant. \code{value} (optional) contains values, normalized to [0,1], to be displayed in color.}

\item{normalize.mass}{logical flag specifying whether the masses should be normalized to the root mass}

\item{monotonic.growth}{optional string. If "down", the masses are forced to increase top-down; if "up", the masses are forced to decrease bottom-up.}

\item{sort}{optional string. If "left", the masses are ordered from left to right; if "right", the masses are sorted from right to left.}

\item{simplify}{logical flag. If TRUE, all the halos (other than the root) with only one progenitor are removed for graphical simplicity}

\item{draw.edges}{logical flag to turn on/off edges}

\item{draw.vertices}{logical flag to turn on/off vertices (halos)}

\item{draw.leaf.vertices}{logical flag to turn on/off the vertices of the leaves}

\item{draw.box}{logical flag to turn on/off a black frame around the box}

\item{root.length}{length of the root branch, extending down from the root halo}

\item{leaf.length}{length of the leaf branches, extending up from the leaves}

\item{style}{integer value to specify the rendering: 1=straight lines, 2=splines, 3=polygon splines with optimal vertex-matching}

\item{scale}{scaling factor to convert masses into line widths}

\item{gamma}{scaling exponent to convert masses into line widths}

\item{min}{minimum linewidth}

\item{vertex.size}{scaling factor of the vertex size}

\item{margin}{thickness of margin surrouding the plot}

\item{add}{logical flag specifying whether the plot should be added to an existing plot}

\item{xlim}{horizontal range of the tree with margin (only needed if add = FALSE)}

\item{ylim}{vertical range of the tree with margin (only needed if add = FALSE)}

\item{xrange}{horizontal range to be displayed}

\item{yrange}{vertical range to be displayed}

\item{col}{color of the tree; or a vector of colors if values have been specified in the tree}

\item{pdf.filename}{optional filename of a PDF file to be saved}

\item{pdf.size}{optional size of the PDF image}
}
\description{
Visualise rooted directional trees, such as the merger trees of halos and galaxies.
}
\examples{

## plot default tree
showtree()

## show a simple custom tree without vertices
tree = list(mass = c(1,0.6,0.1,0.5,0.4), descendant = c(0,1,2,2,1))
showtree(tree)

}
\seealso{
\code{\link{buildtree}}
}
\author{
Danail Obreschkow
}
