% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph2car.R
\name{sph2car}
\alias{sph2car}
\title{Spherical to cartesian coordinate conversion}
\usage{
sph2car(x)
}
\arguments{
\item{x}{3-element or n-by-3 matrix representing the spherical components (r,theta,phi) of n three-dimensional vectors. Here, theta=0...pi is the polar angle measured from the north pole and phi=0...2*pi is the azimuth measured positively from the x-axis (ISO 80000-2:2019 physics convention).}
}
\value{
Returns a 3-element vector or a n-by-3 element matrix representing the Cartesian coordinates (x,y,z)
}
\description{
Convert 3D spherical to Cartesian coordinates
}
\seealso{
\code{\link{car2sph}}
}
\author{
Danail Obreschkow
}
