% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericalY.R
\name{sphericalY}
\alias{sphericalY}
\title{Spherical Harmonics}
\usage{
sphericalY(l, m, x = NULL, theta = NULL, phi = NULL)
}
\arguments{
\item{l}{degree of the spherical harmonic (0,1,2,3)}

\item{m}{sub-order of the spherical harmonic (-l,-l+1,...,+l)}

\item{x}{n-by-3 matrix [1:n,1:3] specifying the 3D coordinates of n points; if x is given, theta and phi cannot be specified.}

\item{theta}{n-vector of azimuth angles from 0 to pi; if theta is given, phi must also be given, but x must not be given.}

\item{phi}{n-vector of longitude angles from 0 to 2*pi; if phi is given, theta must also be given, but x must not be given.}
}
\value{
Returns an n vectors of the spherical harmonics.
}
\description{
Evaluates complex-valued spherical harmonics in the Condon-Shortley phase convention.
}
\author{
Danail Obreschkow
}
