% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{stop}
\alias{stop}
\title{Stop code and produce an error message}
\usage{
stop(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character (and which are pasted together with no separator) or a single condition object.}
}
\value{
None. geterrmessage() returns the last error message, as a character string.
}
\description{
Stops the code and produces an error message, like the stop() function of the base library, but terminating a timer started with \code{\link{tick}}, if needed.
}
\seealso{
\code{\link{tick}} \code{\link{tock}} \code{\link{progress}}
}
\author{
Danail Obreschkow
}
