% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick.R
\name{tick}
\alias{tick}
\title{Start timer}
\usage{
tick(txt = "Start")
}
\arguments{
\item{txt}{custom text}
}
\value{
None
}
\description{
Start timer and write a custom text into the console.
}
\examples{

tick('Sum 10 million random numbers')
x = sum(runif(1e7))
tock()

}
\seealso{
\code{\link{tock}}
}
\author{
Danail Obreschkow
}
