% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tock.R
\name{tock}
\alias{tock}
\title{Stop timer}
\usage{
tock(txt = "", fmt = " (\%.2fs). \%s\\n")
}
\arguments{
\item{txt}{custom text to be displayed.}

\item{fmt}{character vector of format strings. It must contain exactly one \%s as placeholder for \code{txt} and one numerical format, such as \%f or \%e, as placeholder for the time.}
}
\value{
Returns the elapsed time in seconds since calling tick().
}
\description{
Stop timer and write the computation in seconds since the last call of tick().
}
\examples{

tick('Sum 10 million random numbers')
x = sum(runif(1e7))
tock()

}
\seealso{
\code{\link{tick}} \code{\link{progress}} \code{\link{stop}}
}
\author{
Danail Obreschkow
}
