% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparent.R
\name{transparent}
\alias{transparent}
\title{Add transparency to a color}
\usage{
transparent(col, alpha = 0.3)
}
\arguments{
\item{col}{is a color or vector of colors, specified as text (e.g. 'purple') or 7/9-character (e.g. '#A020F0')}

\item{alpha}{tranparency value between 0 and 1}
}
\value{
Returns a 9-character color or vector of 9-character colors.
}
\description{
Add transparency to a color
}
\examples{
col = transparent('purple',0.5)

}
\author{
Danail Obreschkow
}
