% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectornorm.R
\name{vectornorm}
\alias{vectornorm}
\title{Vector norm}
\usage{
vectornorm(x)
}
\arguments{
\item{x}{m-element vector or n-by-m matrix (x[1:n,1:m]) representing n m-element vectors}
}
\value{
Returns a scalar or an n-element vector with the vector norms
}
\description{
Compute the norm of a vector or a matrix of vectors
}
\seealso{
\code{\link{vectornorm}}
}
\author{
Danail Obreschkow
}
