% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yinyangyong.R
\docType{data}
\name{yinyangyong}
\alias{yinyangyong}
\title{Yin-Yang-Yong image}
\format{
An object of class \code{array} of dimension 500 x 500 x 4.
}
\usage{
yinyangyong
}
\value{
None
}
\description{
A data set containing a 500-by-500-by-4 array, representing an image of a Yin-Yang-Yong symbol with transparency. This symbol was 'reinvented' many times. The particular version in this package is the one drawn by the author D. Obreschkow in 1999.
}
\examples{
nplot(pty='s')
rasterImage(yinyangyong,0,0,1,1)

}
\author{
Danail Obreschkow
}
\keyword{datasets}
