% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmplx2col.R
\name{cmplx2col}
\alias{cmplx2col}
\title{Convert complex numbers to color}
\usage{
cmplx2col(z, max = NULL, hue = 0, saturation = 1, gamma = 1)
}
\arguments{
\item{z}{complex number or vector/array of complex numbers}

\item{max}{value of the complex module that corresponds to full brightness; if not given, this is set equal to the maximum complex amplitude in z.}

\item{hue}{hue value of positive reals}

\item{saturation}{saturation value of colors}

\item{gamma}{positive number to adjust the non-linearity of the color scale (1=linear)}
}
\value{
Returns a single color or vector/array of colors
}
\description{
Converts a complex number (or a vector/array thereof) into a color-string, such that brightness represents the complex amplitude and hue represents the complex phase.
}
\author{
Danail Obreschkow
}
