% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst.R
\docType{data}
\name{cst}
\alias{cst}
\title{Scientific constants}
\format{An object of class \code{list} of length 15.}
\usage{
cst
}
\description{
The list \code{cst} contains a lot of useful scientific constants in SI units, mainly for astrophysics.
}
\author{
Danail Obreschkow
}
\keyword{datasets}
