% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.env.R
\docType{data}
\name{.cooltools.env}
\alias{.cooltools.env}
\title{Package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.cooltools.env
}
\description{
Environment used to store global variables in the package.
}
\author{
Danail Obreschkow
}
\keyword{datasets}
