% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertia.R
\name{inertia}
\alias{inertia}
\title{Inertia tensor}
\usage{
inertia(x, m = 1)
}
\arguments{
\item{x}{n-by-3 matrix (x[1:n,1:3]) specifying the 3D Cartesian coordinates of n points}

\item{m}{n-vector with point masses, or single scalar giving a uniform mass for all points (default is unity)}
}
\value{
Returns a 3-by-3 symmetric matrix
}
\description{
Computes the symmetric tensor of moments of inertia
}
\seealso{
\code{\link{quadrupole}}, \code{\link{moments}}
}
\author{
Danail Obreschkow
}
