% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Second moment tensor}
\usage{
moments(x, m = 1)
}
\arguments{
\item{x}{n-by-3 matrix (x[1:n,1:3]) specifying the 3D Cartesian coordinates of n points}

\item{m}{n-vector with point masses, or single scalar giving a uniform mass for all points (default is unity)}
}
\value{
Returns a 3-by-3 symmetric matrix
}
\description{
Compute the tensor of second moments of a set of point masses
}
\seealso{
\code{\link{inertia}}, \code{\link{quadrupole}}
}
\author{
Danail Obreschkow
}
