% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplot.R
\name{nplot}
\alias{nplot}
\title{Make empty plot area}
\usage{
nplot(
  xlim = c(0, 1),
  ylim = c(0, 1),
  xlab = "",
  ylab = "",
  xaxs = "i",
  yaxs = "i",
  xaxt = "n",
  yaxt = "n",
  bty = "n",
  ...
)
}
\arguments{
\item{xlim, ylim}{vectors with plotting limits.}

\item{xlab, ylab}{horizontal and vertical labels.}

\item{xaxs, yaxs}{style of the axis interval (see \code{\link[graphics]{par}}).}

\item{xaxt, yaxt}{character which specifies the x axis type (see \code{\link[graphics]{par}}).}

\item{bty}{character specifying the border type (see \code{\link[graphics]{par}}).}

\item{...}{additional arguments used by \code{\link[graphics]{plot}}}
}
\value{
None
}
\description{
Open an empty plot
}
\author{
Danail Obreschkow
}
