% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrupole.R
\name{quadrupole}
\alias{quadrupole}
\title{Quadrupole tensor}
\usage{
quadrupole(x, m = 1)
}
\arguments{
\item{x}{n-by-3 matrix (x[1:n,1:3]) specifying the 3D cartesian coordinates of n points}

\item{m}{n-vector with point masses, or signle scalar giving a uniform mass for all points (defaults is unity)}
}
\value{
Returns a 3-by-3 symmetric matrix
}
\description{
Compute the trace-free quadrupole tensor of a set of point masses
}
\author{
Danail Obreschkow
}
