% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Suppress in-routine output}
\usage{
quiet(x)
}
\arguments{
\item{x}{routine to be called}
}
\value{
Returns whatever the called routine returns in invisible form.
}
\description{
Runs any routine or command while supressing in-routine console output
}
\examples{
# Test function
test = function(x) {
  cat('This routine is likes to talk a lot!\n')
  return(x^2)
}

# Standard call call:
y = test(5)
print(y)

# Quiet call:
y = quiet(test(6))
print(y)

}
\author{
Danail Obreschkow
}
