% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readhdf5.R
\name{readhdf5}
\alias{readhdf5}
\title{Read data from an HDF5 file}
\usage{
readhdf5(file)
}
\arguments{
\item{file}{A character string specifying the path to the input HDF5 file.}
}
\value{
A nested list representing the contents of the HDF5 file. Groups are represented as lists, datasets are represented by their data, and attributes are included either as the main data (if no other data is present) or under the `_attributes` key for groups.
}
\description{
Recursively reads data from an HDF5 file, preserving the structure of groups, datasets, and attributes. The function supports groups and datasets containing only attributes, treating these attributes as the main data for those objects (e.g. as used in the Header group of HDF5 outputs of the Gadget simulation code).
}
\details{
This function reads the contents of an HDF5 file into a nested list structure in R. Groups and datasets are recursively parsed, preserving the original hierarchy.
If a group or dataset contains only attributes and no other data, these attributes are returned as the main data for that object. Attributes for objects with actual data are ignored, except for groups, where they are stored under the `_attributes` key.
}
\examples{

\dontrun{
# make test data
input = list(
  group1 = list(
    dataset1 = matrix(1:10, nrow = 2),
    dataset2 = c("A", "B", "C")
  ),
  group2 = list(
    dataset3 = c(3.14, 2.71),
    subgroup1 = list(
      dataset3 = array(runif(8), dim = c(2, 2, 2))
    )
  ),
  group3 = list(id64 = bit64::as.integer64(123487918235335756), id32 = 5756)
)

# write list to HDF5 file
writehdf5(example_data, "example_data.h5")

# read HDF5 into a new list
output = readhdf5(filename)

# Test if input and output lists are identical
print(all.equal(input,output))
}

}
\seealso{
\code{\link[hdf5r]{H5File}}, \code{\link{writehdf5}}
}
