% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readhdf5.R
\name{readhdf5}
\alias{readhdf5}
\title{Read data from an HDF5 file}
\usage{
readhdf5(file, subtree = NULL, group.attr.as.data = FALSE)
}
\arguments{
\item{file}{Character string specifying the file name of the input HDF5 file.}

\item{subtree}{Optional hierarchical list, specifying groups and datasets to read from the HDF5 file. For example, to read group1-dataset3, group5-dataset7, group5-dataset8, and group7-subgroup3-dataset9, use \code{subtree = list(group1 = list(dataset3 = ""), group5 = list(dataset7 = "", dataset8 = ""), group7 = list(subgroup3 = list(dataset9 = "")))}.
If \code{subtree} is NULL (default), the entire file hierarchy is read.}

\item{group.attr.as.data}{logical flag. If \code{TRUE}, group attributes are converted to datasets. This is useful, for example, for Gadget simulation outputs in HDF5 format, where the simulation parameters are stored as attributes in the `Header' group.}
}
\value{
A nested list representing the contents of the HDF5 file. Groups are represented as lists, datasets are represented by their data, and attributes are included either as the main data (if no other data is present) or under the `_attributes` key for groups.
}
\description{
Reads data from an HDF5 file, including attributes, into a nested list that preserves the hierarchical structure of the HDF5 data.
}
\details{
This function, built on the \code{hdf5r} package, reads the contents of an HDF5 file into a nested list structure in R. Groups and datasets are recursively parsed, preserving the original hierarchy.
}
\seealso{
\code{\link[hdf5r]{H5File}}, \code{\link{writehdf5}}
}
