% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalarproduct.R
\name{scalarproduct}
\alias{scalarproduct}
\title{Scalar product}
\usage{
scalarproduct(x, y)
}
\arguments{
\item{x, y}{d-element vectors or n-by-d matrices representing n d-element vectors}
}
\value{
Returns a scalar or a n-element vector with the scalar products
}
\description{
Compute scalar product of two vectors
}
\seealso{
\code{\link{vectorproduct}}
}
\author{
Danail Obreschkow
}
