% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tock.R
\name{tock}
\alias{tock}
\title{Stop timer}
\usage{
tock(txt = "")
}
\arguments{
\item{txt}{optional custom text to be displayed}
}
\value{
None
}
\description{
Stop timer and write the computation in seconds since the last call of tick().
}
\examples{

tick('Sum 10 million random numbers')
x = sum(runif(1e7))
tock()

}
\seealso{
\code{\link{tick}}
}
\author{
Danail Obreschkow
}
