% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitvector.R
\name{unitvector}
\alias{unitvector}
\title{Normalize vectors to unit length}
\usage{
unitvector(x)
}
\arguments{
\item{x}{m-element vector or n-by-m matrix \code{x[1:n,1:m]} representing n m-element vectors}
}
\value{
Returns a data type identical to \code{x}, but with normalised vectors. Zero-vectors are returned for vectors of length zero.
}
\description{
Compute the unit vectors for for a matrix of non-normalised vectors
}
\seealso{
\code{\link{vectornorm}}
}
\author{
Danail Obreschkow
}
