% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorproduct.R
\name{vectorproduct}
\alias{vectorproduct}
\title{Vector product}
\usage{
vectorproduct(x, y, normalize = FALSE)
}
\arguments{
\item{x, y}{3-element vectors or n-by-3 matrices representing n 3-element vectors}

\item{normalize}{logical flag; if set to TRUE the cross-product(s) is/are automatically normalized}
}
\value{
Returns a 3-element vector or a n-by-3 element matrix with the cross products
}
\description{
Compute cross product of two 3-element vectors
}
\seealso{
\code{\link{scalarproduct}}, \code{\link{vectornorm}} and \code{\link{unitvector}}
}
\author{
Danail Obreschkow
}
