% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writehdf5.R
\name{writehdf5}
\alias{writehdf5}
\title{Write data to an HDF5 file}
\usage{
writehdf5(data, file, level = 6, overwrite = TRUE)
}
\arguments{
\item{data}{A list containing the data to be written. Nested lists are interpreted as groups within the HDF5 file. If an element is `NULL`, it creates an empty dataset that can hold only attributes.}

\item{file}{A character string specifying the path to the output HDF5 file.}

\item{level}{An integer specifying the compression level for datasets, typically between 0 (no compression) and 9 (maximum compression). Not all dataset types support compression.}

\item{overwrite}{A logical value indicating whether to overwrite an existing file.}
}
\value{
This function does not return any value. It writes data directly to the specified HDF5 file.
}
\description{
Recursively writes data to an HDF5 file, creating groups and datasets as needed. The routine supports writing lists, 64-bit integers, character data, and numeric arrays.
}
\details{
The function creates the necessary subdirectories for the output file if they do not exist.
It uses 64-bit integer format (\code{H5T_NATIVE_INT64}) for \code{integer64} data and automatically detects types for character and numeric data.
Attributes can be added to groups or datasets that contain no other data, allowing for flexible metadata storage.
}
\examples{

\dontrun{
# Create example data
input <- list(
  group1 = list(
    dataset1 = matrix(1:10, nrow = 2),
    dataset2 = c("A", "B", "C")
  ),
  group2 = list(
    dataset3 = c(3.14, 2.71),
    subgroup1 = list(
      dataset3 = array(runif(8), dim = c(2, 2, 2))
    )
  ),
  group3 = list(id32 = 5756, id64 = bit64::as.integer64(123487918235335756)),
  group4 = 4
)
# Write list to HDF5 file
writehdf5(input, "example_data.h5")
# Read HDF5 into a new list
output <- readhdf5("example_data.h5")
# Test if input and output lists are identical
print(all.equal(input, output))
}

}
\seealso{
\code{\link[hdf5r]{H5File}}, \code{\link{readhdf5}}
}
